requirements.txt
fastapi
uvicorn[standard]
requests
gradio==4.28.3
ollama
python-dotenv